refactor(provider-sdk)!: u32 scope and channel; drop tungstenite
`scope` and `channel` become `u32`, and the varint codec with them.
Header now costs three bytes minimum and eleven at worst, against
twenty-one for `u64`. Four billion scopes per connection and four
billion channels per scope is not a ceiling anyone reaches.

`tungstenite` is gone, along with the two `Message` conversions and
`FrameError::NotBinary`. Back to five dependencies.

Removing it was the right call and the reason is worth recording,
because someone will propose adding it back: there is no single
`Message` type to convert against. axum defines its OWN
(`axum/src/extract/ws.rs:772`) rather than re-exporting tungstenite's,
converting internally via `into_tungstenite()`. It also pins
tokio-tungstenite 0.29 while the latest tungstenite is 0.30 — so the
impls served neither axum, nor the laboratory and SDK on 0.24, nor
anything but a consumer on exactly 0.30. Four transitive crates for a
two-line convenience nothing in this repo could use.

What works instead, with no dependency and no version coupling: every
`Message::Binary` carries `bytes::Bytes`, which derefs to `&[u8]`. So
`decode(&bytes)` and `Message::Binary(frame.encode().into())` compile
against any of them. That is now the module doc's own example.

+++ objectiveai-provider-sdk/src/frame/client.rs
@@ -19,17 +19,17 @@ pub enum ClientFrame<'a> {
     /// begun.
     Ack {
         /// The scope the server minted.
-        scope: u64,
+        scope: u32,
         /// The channel of the server request being answered.
-        channel: u64,
+        channel: u32,
     },
     /// Type `1`. One piece of the answer. There may be any number,
     /// including none.
     Body {
         /// The scope the server minted.
-        scope: u64,
+        scope: u32,
         /// The channel of the server request being answered.
-        channel: u64,
+        channel: u32,
         /// The body bytes.
         payload: &'a [u8],
     },
@@ -37,9 +37,9 @@ pub enum ClientFrame<'a> {
     /// Nothing follows on it.
     Finish {
         /// The scope the server minted.
-        scope: u64,
+        scope: u32,
         /// The channel of the server request being answered.
-        channel: u64,
+        channel: u32,
     },
     /// Type `3`. A new request, opening a scope.
     ///
@@ -53,13 +53,8 @@ pub enum ClientFrame<'a> {
 }
 
 impl<'a> ClientFrame<'a> {
-    /// Decode one frame. The payload borrows from `bytes`.
-    ///
-    /// `bytes` is the WebSocket message's binary payload. Nothing here
-    /// names a WebSocket type: `axum`'s `Message::Binary` and
-    /// tungstenite's both carry `Bytes`, which derefs to `&[u8]`, so
-    /// this takes the one thing they agree on and the SDK depends on
-    /// neither.
+    /// Decode one frame from a WebSocket message's binary payload.
+    /// The payload borrows from `bytes`.
     pub fn decode(bytes: &'a [u8]) -> Result<Self, FrameError> {
         let (r#type, scope, channel, payload) = super::split_header(bytes)?;
         Ok(match r#type {
@@ -80,7 +75,7 @@ impl<'a> ClientFrame<'a> {
     ///
     /// The one place the mapping lives, so the length and the writer
     /// cannot disagree about it.
-    fn parts(&self) -> (u8, u64, u64, &'a [u8]) {
+    fn parts(&self) -> (u8, u32, u32, &'a [u8]) {
         match *self {
             ClientFrame::Ack { scope, channel } => (0, scope, channel, &[]),
             ClientFrame::Body {
